Slightly better color reporting
Now reports the color of the LED as the string for the measure. I am
trying to get it working for stuff I dont know what the color should be
but I can not get it to report back what the color is.
I am also trying to get init and uninit of the chroma instances to play
nice but it seems the check for if it is initialized does not work/I am
using it wrong. Uninit also seems to take about a second to finish so if
I setup my effect before then I run into issues. Looking into fixes that
wont cause rainmeter to lock up.