Full keyboard support
Big update there are now two tester files that are slightly different.
They are going to become more different in the future and one will
become a showoff file with the other being comprehensive examples.

Keyboard support is ongoing, I think I am going to rewrite some more as
the current setup seems to not being referencing the keys to set but I
will need to get a tester to confirm it.
Also fixed is the refresh bug, with any luck this implementation in the
future will allow for fixing the gaming and screen off bugs. As well as
allow support for going back to previous lights when a skin is unloaded.