Stated process of adding corsair support
Given the issues I am having with Razer's API and the requests for
corsair support I am now working on adding support for corsair devices.
I imagine testing of this is going to be slow so I think my approach is
going to be implement everything the same as razer and then send to
testers.